requirements.txt
python-dotenv
werkzeug==2.1.2
Flask~=2.1.2
flask-restx==0.5.1
click<8.1.0
pyyaml==6.0.0
ruamel.yaml==0.17.21
itsdangerous
Jinja2
MarkupSafe
uWSGI
gtts
pysqlite3
redis==4.3.4
Flask-Caching==2.0.1
Flask-APScheduler
fakeyou
sqlitedict
Flask-Limiter
uwsgi-tools
pydub
tqdm
